requirements.txt
openai
chainlit
chromadb
litellm
python-dotenv
numpy
pytest
pydantic